refact: 구독 로직 분리
