+++ components/nav.js
@@ -6,18 +6,13 @@ import {
   myList,
   getMediaLength,
   getMyListLength,
-  categoryData,
-  addMyData,
-  removeMyDataById,
   mediaData,
-  myData,
 } from "../resources/data.js";
 import state from "../list/state.js";
-import store from "../utils/stoageManager.js";
+import { generateSubscribe, generateUnsubscribe } from "./subscribe.js";
 
 var newsInterval; //왜 전역변수로 설정해야 작동하는가??
 var myInterval;
-let needRemove = [];
 
 /**
  * Nav목록으로 네비게이션 바를 container의 child로 생성 후 초기화
@@ -98,45 +93,9 @@ function generateNavForNewsList() {
 }
 
 /**
- * 구독 버튼에 이벤트 추가
- */
-function generateSubscribe() {
-  const contentTitle = document.querySelector(".content_title");
-  contentTitle.lastChild.remove();
-  const sub = document.createElement("button");
-  sub.classList.add("subscribe");
-  contentTitle.appendChild(sub);
-
-  needRemove.forEach((element) => removeMediaFromMyList(element));
-  needRemove = [];
-
-  sub.innerHTML = "+ 구독하기";
-  sub.addEventListener("click", addMediaToMyList);
-}
-
-/**
- * 구독 버튼 로직
- */
-function addMediaToMyList() {
-  store.addItemToSet(
-    "myList",
-    categoryData[categoryList[state.currentCategoryIndex]][
-      state.currentMediaIndex
-    ]
-  );
-  addMyData(
-    mediaData[
-      categoryData[categoryList[state.currentCategoryIndex]][
-        state.currentMediaIndex
-      ]
-    ]
-  );
-}
-
-/**
  * 구독 newsList를 생성하는 함수
  */
-function generateNavForMyList() {
+export function generateNavForMyList() {
   updateMyNewsList(state.currentCategoryIndex);
 
   generateUnsubscribe();
@@ -146,34 +105,6 @@ function generateNavForMyList() {
   setupNavElements(navElementNodes, startMyNewsInterval, updateMyMedia);
 
   startMyNewsInterval();
-}
-
-/**
- * 구독해지 버튼에 이벤트 추가
- */
-function generateUnsubscribe() {
-  const contentTitle = document.querySelector(".content_title");
-  contentTitle.lastChild.remove();
-  const sub = document.createElement("button");
-  sub.classList.add("subscribe");
-  contentTitle.appendChild(sub);
-
-  sub.innerHTML = "구독 해지";
-  sub.addEventListener("click", () =>
-    saveRemoveList(myList[state.currentCategoryIndex])
-  );
-}
-
-/**
- * 구독해지 버튼 로직
- */
-function removeMediaFromMyList(id) {
-  store.removeItemFromSet("myList", id);
-  removeMyDataById(mediaData[id].media);
-}
-
-function saveRemoveList(id) {
-  needRemove.push(id);
 }
 
 /**
